Silence codecvt deprecation warnings when building with NDK 28 (#2)
According to these sources:
https://stackoverflow.com/questions/42946335/deprecated-header-codecvt-replacement
https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2017/p0618r0.html

We should not be in a hurry to replace the deprecated standard library APIs, and we're actually not using the affected functions anyway, as we don't have wstrings that need to be passed across the JNI boundaries.

For now we're just going to silence the warnings.